requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.3.2
PyMuPDF==1.23.26
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
python-dotenv==1.0.1
ratelimit==2.2.1